Important change
The previous code did not take into account if the @needle string was found
continuosly. Eg, If the @haystack string whas "hello world" and @needle was
"lord" it were going to print "ello world" meaning that @needle was found, and
that is not correct.

To solve that, an Else If was added in the line 35, wich includes a reset to
several variables.